chore: Move chat to providers crate
