--- chat/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "muchat_chat"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-tokio = { version = "1.28", features = ["full"] }
-tokio-tungstenite = "0.19"
-futures-util = "0.3"
-futures = "0.3.30"
-url = "2.4"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-chrono = { version = "0.4", features = ["serde"] }
