+++ providers/src/chat/error.rs
