+++ providers/src/chat/response.rs
